Added API calls to game board
Added functionality to interface game with the API. "New Game" now
makes a call to generate a game object and store it in store.js.
Each move is pushed to the API endpoint to update the API's record of
the game. Also added curl scripts

Next Steps: refactor js to less of a mess. Also begin surfacing current
console logs.